requirements.txt
amzqr==0.0.1
anyio==4.0.0
asn1crypto==1.5.1
certifi==2023.7.22
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
imageio==2.31.5
numpy==1.26.1
pg8000==1.30.2
Pillow==10.1.0
python-dateutil==2.8.2
python-telegram-bot==20.6
scramp==1.4.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
wheel==0.41.2
